requirements.txt
cffi==1.16.0
colorama==0.4.6
contourpy==1.2.1
convert-files==0.1
cycler==0.12.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fonttools==4.52.1
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib_resources==6.4.0
iniconfig==2.0.0
jcw-utilities==0.1
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.0
networkx==3.2.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pillow_heif==0.16.0
pluggy==1.5.0
psutil==5.9.8
pycparser==2.22
pyheif==0.7.1
pyparsing==3.1.2
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-image==0.22.0
scipy==1.13.1
six==1.16.0
tifffile==2024.5.22
tomli==2.0.1
tqdm==4.66.4
tzdata==2024.1
vtk==9.3.0
zipp==3.19.0